Add babel-plugin-transform to fix regenerator error
