+++ package.json
@@ -16,7 +16,6 @@
   "author": "Andela Simulations Programme",
   "license": "MIT",
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
     "bcrypt": "^3.0.3",
     "chai": "^4.2.0",
     "chai-http": "^4.2.1",
@@ -44,6 +43,8 @@
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-watch": "^2.0.8",
     "coveralls": "^3.0.2",
